+++ src/components/Workspace.tsx
@@ -4,6 +4,8 @@ import { useStoreActions, useStoreState } from '../stores/Hooks';
 import { ControlHandler } from './Blocks/ControlHandler';
 import { MeshGradient } from './Misc/MeshGradient';
 import { LavaLampBackground } from './Misc/LavaLampBackground';
+import { StarfieldBackground } from './Misc/StarfieldBackground';
+import { GalaxyBackground } from './Misc/GalaxyBackground';
 import Moveable, {
 	type OnDrag,
 	type OnResize,
@@ -222,8 +224,36 @@ export const Workspace: React.FC<Props> = ({ reference }) => {
 										blurSpread * 2
 									}
 								/>
-							) : (
+							) : currentWorkspace?.workspaceDynamicType === 'lava' ? (
 								<LavaLampBackground
+									colors={currentWorkspace?.workspaceDynamicSettings.colors}
+									blur={currentWorkspace?.workspaceBlur}
+									seed={currentWorkspace?.workspaceDynamicSettings.seed}
+									width={
+										parseInt(currentWorkspace?.workspaceWidth || '512') +
+										blurSpread * 2
+									}
+									height={
+										parseInt(currentWorkspace?.workspaceHeight || '512') +
+										blurSpread * 2
+									}
+								/>
+							) : currentWorkspace?.workspaceDynamicType === 'starfield' ? (
+								<StarfieldBackground
+									colors={currentWorkspace?.workspaceDynamicSettings.colors}
+									blur={currentWorkspace?.workspaceBlur}
+									seed={currentWorkspace?.workspaceDynamicSettings.seed}
+									width={
+										parseInt(currentWorkspace?.workspaceWidth || '512') +
+										blurSpread * 2
+									}
+									height={
+										parseInt(currentWorkspace?.workspaceHeight || '512') +
+										blurSpread * 2
+									}
+								/>
+							) : (
+								<GalaxyBackground
 									colors={currentWorkspace?.workspaceDynamicSettings.colors}
 									blur={currentWorkspace?.workspaceBlur}
 									seed={currentWorkspace?.workspaceDynamicSettings.seed}